Move set_logger_verbosity and setup_outputdir to common (#1403)
* move set_logger_verbosity and setup_outputdir to common

* fix

* fix

* fix

* fix

* logger refactor
