+++ vision/src/autogluon/vision/predictor/predictor.py
@@ -14,11 +14,11 @@ try:
 except ImportError:
     timm = None
 
+from autogluon.common.utils.log_utils import set_logger_verbosity, verbosity2loglevel
 from autogluon.common.utils.multiprocessing_utils import is_fork_enabled
 from autogluon.core.constants import MULTICLASS, BINARY, REGRESSION
 from autogluon.core.data.label_cleaner import LabelCleaner
-from autogluon.core.utils import set_logger_verbosity
-from autogluon.core.utils import verbosity2loglevel, get_gpu_count_all
+from autogluon.core.utils import get_gpu_count_all
 from autogluon.core.utils.utils import generate_train_test_split
 
 from ..configs.presets_configs import unpack, _check_gpu_memory_presets
@@ -27,7 +27,7 @@ from ..utils import sanitize_batch_size
 __all__ = ['ImagePredictor']
 
 
-logger = logging.getLogger()  # return root logger
+logger = logging.getLogger(__name__)  # return autogluon root logger
 
 
 class ImagePredictor(object):
@@ -264,7 +264,7 @@ class ImagePredictor(object):
         tuning_data = copy.deepcopy(tuning_data)
 
         log_level = verbosity2loglevel(self._verbosity)
-        set_logger_verbosity(self._verbosity, logger=logger)
+        set_logger_verbosity(self._verbosity)
         if presets:
             if not isinstance(presets, list):
                 presets = [presets]
